Working Protest object
This commit provides a functioning Protest object. The Protest object is the heart of
where the data processing will happen to make sure that the html in the docket listing
is converted into data.

Changed the test directory to distinguish between the tests and the data that the tests
rely upon.