Add full text opinions
This commit adds an Opinion class, which gets the full text of a decision
and the summary, and adds them as part of the Protest data. Some clean up
of tests and the classes are needed to accommodate this.

The run.py script is also changed to allow for the dump of the data into
the sqlite db and the protests.csv.